org: code block evaluation wip
