+++ scripts/py_haxorg/py_haxorg/pyhaxorg.pyi
@@ -556,7 +556,7 @@ class OrgCodeEvalInputResultHandling(Enum):
     Prepend = 6
 
 class OrgCodeEvalInput:
-    def __init__(self, blockAttrs: AttrGroup, tangledCode: str, resultType: OrgCodeEvalInputResultType, resultFormat: OrgCodeEvalInputResultFormat, resultHandling: OrgCodeEvalInputResultHandling) -> None: ...
+    def __init__(self, blockAttrs: AttrGroup, tangledCode: str, resultType: OrgCodeEvalInputResultType, resultFormat: OrgCodeEvalInputResultFormat, resultHandling: OrgCodeEvalInputResultHandling, language: str) -> None: ...
     def operator==(self, other: OrgCodeEvalInput) -> bool: ...
     def __repr__(self) -> str: ...
     def __getattr__(self, name: str) -> object: ...
@@ -565,6 +565,7 @@ class OrgCodeEvalInput:
     resultType: OrgCodeEvalInputResultType
     resultFormat: OrgCodeEvalInputResultFormat
     resultHandling: OrgCodeEvalInputResultHandling
+    language: str
 
 class OrgCodeEvalOutput:
     def __init__(self, stdout: str, stderr: str, code: int) -> None: ...
@@ -1726,7 +1727,7 @@ class BlockCodeEvalResult(Block):
 class BlockCode(Block):
     def __init__(self, lang: Optional[str], switches: List[BlockCodeSwitch], exports: BlockCodeExports, result: List[BlockCodeEvalResult], lines: List[BlockCodeLine], cache: bool, eval: bool, noweb: bool, hlines: bool, tangle: bool, attrs: Optional[AttrGroup], attached: List[Org]) -> None: ...
     def getVariable(self, varname: str) -> Optional[AttrValue]: ...
-    def getCodeForEvaluation(self) -> OrgCodeEvalInputResultType: ...
+    def getCodeForEvaluation(self) -> OrgCodeEvalInput: ...
     def getAttrs(self, key: Optional[str]) -> List[AttrValue]: ...
     def getFirstAttr(self, kind: str) -> Optional[AttrValue]: ...
     def getAttached(self, kind: Optional[str]) -> List[Org]: ...
